Split DSL types, rename infer help, and tighten param config
- Move display/inferred-command DSL types into command-dsl-types.ts;
  re-export from infer-commands so consumers are unaffected
- Mark generated-command-dsl-types.ts as linguist-generated with -diff
  so the regen noise stops dominating PRs
- Split InferredCommandParamConfig into argument vs option variants:
  required is boolean-only on arguments and forbidden on options;
  inferFrom drives project/team context lookup
- Restore the bare --help fast path in vc.js; switch the inference
  help trigger to --infer-help so --help stays unambiguous
- Trim runInferredCommand call sites in index.ts: derive cwd/scope/
  team/api from the Client instead of duplicating flag extraction
