+++ .gitattributes
@@ -14,3 +14,6 @@ go.mod text eol=lf
 **/test/fixtures/**/git/**/* linguist-generated
 
 .github/workflows/*.lock.yml linguist-generated=true merge=ours
+
+# Generated by scripts/generate-openapi-command-dsl-types.mjs
+packages/cli/src/util/openapi/generated-command-dsl-types.ts linguist-generated=true -diff
